requirements.txt
# Agent SDK + LLM
aws-strands-sdk>=0.1.0
boto3
python-dotenv
pydantic

# Web
fastapi
uvicorn[standard]
httpx
mangum

# Testing
pytest
pytest-cov

# Optional Dev Tools
black
ruff
